updated scripts to be network specific
